requirements.txt
streamlit==0.89.0
PyPDF2==1.26.0
pandas==1.3.3
langchain==0.1.0
langchain-google-genai==0.1.0
google-generativeai==0.6.0
python-dotenv==0.19.1
Pillow==9.0.0
